Benchmark code for fault tolerance (#23)
* Refactor: Decouple dns-store and dns serving

Decouple dns store, which is essentially a k-v store built upon Raft protocol,
  and dns serving logic, like handling non-recursive and recursive dns queries
  using dns store.

* Tools: add bash file to do cleanup after run local test. Run './tools/cleanup.sh'

* Refactor: Correct typo

* Tools: update makefile to do cleanup

* Feat: Remote start and kill dns server

Add support for remote starting and killing dns server.
Run `go build -o remote` and `./remote` on each node of dns server,
  it will listen to port 8090.
Send '/start' or 'kill' to the port of server to start or kill the dns server.

* Feat(remote start): Support argument start from HTTP request body

* benchmark script

* throughput basis

* refractor

* rename

* read/write ratio option

* fixed too-many-write-to-same-domain

* added populating db

* added benchmark instruction

* Update usage.md

Added "-ip" field

* Benchmark(fault tolerance): add basic client code for fault tolerance test

* Benchmark(fault tolerance): add support for presenting throughput at regular intervals

* Benchmark(fault tolerance): full support for killing and restarting one dns store cluster node, w/o hash server

* Benchmark(fault tolerance): full support for with hash server bench test

* Benchmark(fault tolerance): refine benchmark

+++ dns/tools/remote_shutdown.go
@@ -11,8 +11,12 @@ import (
 )
 
 var cmd *exec.Cmd
+var running bool
 
 func start_dnsserver(w http.ResponseWriter, req *http.Request) {
+	if running {
+		return
+	}
 	body, err := ioutil.ReadAll(req.Body)
 	if err != nil {
 		log.Printf("Cannot read /start body: %v\n", err)
@@ -29,20 +33,26 @@ func start_dnsserver(w http.ResponseWriter, req *http.Request) {
 	if err := cmd.Start(); err != nil {
 		log.Fatal(err)
 	}
-	log.Print("Starting running dns_server with arguement %s", argvstr)
-	fmt.Fprintf(w, "Starting running dns_server with argument %s", argvstr)
+	running = true
+	log.Print("Starting running dns_server with arguement %s\n", argvstr)
+	fmt.Fprintf(w, "Starting running dns_server with argument %s\n", argvstr)
 }
 
 func kill_dnsserver(w http.ResponseWriter, req *http.Request) {
+	if !running {
+		return
+	}
 	if err := cmd.Process.Kill(); err != nil {
 		log.Fatal(err)
 	}
 	// syscall.Kill(cmd.Process.Pid, syscall.SIGINT)
-	log.Print("Killed the running dns_server")
-	fmt.Fprintf(w, "Killed the running dns_server")
+	running = false
+	log.Print("Killed the running dns_server\n")
+	fmt.Fprintf(w, "Killed the running dns_server\n")
 }
 
 func main() {
+	running = false
 	http.HandleFunc("/start", start_dnsserver)
 	http.HandleFunc("/kill", kill_dnsserver)
 	http.ListenAndServe(":8090", nil)
